// frontend/src/api/generated/models/insightSeverity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type InsightSeverity = typeof InsightSeverity[keyof typeof InsightSeverity];


export const InsightSeverity = {
  Info: 'Info',
  Notice: 'Notice',
  Attention: 'Attention',
} as const;